refactor to util method
